+++ src/main/kotlin/in/kkkev/jjidea/jj/CommitId.kt
@@ -1,5 +1,5 @@
 package `in`.kkkev.jjidea.jj
 
-class CommitId(full: String, short: String? = null) : ShortenableImpl(full, short), Revision {
+class CommitId(full: String, short: String? = null) : ShortenableImpl(full, short), Revision, Shortenable {
     override val short get() = super<ShortenableImpl>.short
 }